requirements.txt
faust-streaming==0.8.8
confluent-kafka==1.9.2
rdflib>=6.2.0
aiobotocore==2.4.0
requests==2.28.2
fastavro==1.7.0 
pandas<=2.0.3